Extract MemberLink to its own file
This step is in preparation of testing MemberLink with scraper_test
